refactor: service 상위 디렉토리명 변경
- admin -> manage
